requirements.txt
functions-framework==3.*
polars
pandas
scikit-learn
optuna
lightgbm
xgboost
catboost
supabase
tqdm
numpy
selenium
marimo
google-cloud-secret-manager
webdriver-manager
pyarrow
shap
catboost
tabpfn-client
tabpfn